Anthropic AI interface for chat completion
